assert {imotionsApi}	R Documentation

Stop if a condition is not fulfilled and print an error message.

Description

Stop if a condition is not fulfilled and print an error message.

Usage

assert(condition, message)

Arguments

condition
R expressions which should evaluate to TRUE.

message
An error message.

[Package imotionsApi version 2.0.0 Index]