assertClass {imotionsApi}	R Documentation

Stop if an object is not from the good class and print an error message.

Description

Stop if an object is not from the good class and print an error message.

Usage

assertClass(object, className, message)

Arguments

object
An object which class needs to be checked.

className
A string indicating the class to check.

message
An error message.

[Package imotionsApi version 2.0.0 Index]