createDataHeader {imotionsApi}	R Documentation

Create data header for the file that needs to be uploaded.

Description

Create data header for the file that needs to be uploaded.

Usage

createDataHeader(params, sensorName, scriptName, fileDependency = NULL)

Arguments

params
The list of parameters provided to the script - specific parameters value will be stored as metadata.

sensorName
The name of the new sensor you would like to create.

scriptName
The name of the script used to produce these signals/metrics.

[Package imotionsApi version 2.0.0 Index]