createImObject {imotionsApi}	R Documentation

Create an iMotions object.

Description

Based on a data.table and the type of object wanted, create an imObject or imObjectList.

Usage

createImObject(data, typeOfObject)

Arguments

data
a list/data.frame that needs to be changed into an iMotions object.

typeOfObject
a string that specifies the type of object wanted ("Stimulus","Respondent", "Segment", "Sensor", "Interval").

Value

An iMotions object of the specified type (data.table).

Examples

## Not run: 
object <- list(name = "example")
respondentsObject <- imotionsApi:::createImObject(object, "Respondent")
stimuliObject <- imotionsApi:::createImObject(object, "Stimulus")
segmentsObject <- imotionsApi:::createImObject(object, "Segment")
sensorsObject <- imotionsApi:::createImObject(object, "Sensor")

## End(Not run)

[Package imotionsApi version 2.0.0 Index]