createSignalsMetadata {imotionsApi}	R Documentation

Create metadata for the file that needs to be uploaded.

Description

Create metadata for the file that needs to be uploaded.

Usage

createSignalsMetadata(data, metadata = NULL)

Arguments

data
A data.table containing the signals/metrics to upload (imData object are also accepted).

metadata
Optional - a data.table with metadata information. Column names will be converted to metadata headers and there must be a row corresponding to each data column.

[Package imotionsApi version 2.0.0 Index]