getIntervalsForStimuli {imotionsApi}	R Documentation

Get time intervals for stimuli.

Description

Return the start, end and name of each stimulus.

Usage

getIntervalsForStimuli(study, sensors)

Arguments

study
An imStudy object as returned from imStudy.

sensors
An imSensorList object as returned from getRespondentSensors.

Value

An imIntervalList object (data.table) composed of the start, end and name of each stimulus.

[Package imotionsApi version 2.0.0 Index]