getJSON {imotionsApi}	R Documentation

Retrieve a JSON file from the indicated path/url - optionally allow the user to add more information about to know when an error occurred.

Description

Retrieve a JSON file from the indicated path/url - optionally allow the user to add more information about to know when an error occurred.

Usage

getJSON(connection, url, message = NULL)

Arguments

connection
An imConnection object as returned from imConnection.

url
The url/path where the JSON file is located.

message
Optional: a short message indicating which steps are getting performed to get a more indicative error message.

Value

The retrieved JSON file.

[Package imotionsApi version 2.0.0 Index]