getRespondentSensors {imotionsApi}	R Documentation

Get all sensors available for a given respondent.

Description

Retrieves detailed information about sensors available for a given respondent.

Usage

getRespondentSensors(study, respondent)

Arguments

study
An imStudy object as returned from imStudy.

respondent
An imRespondent object as returned from getRespondents.

Value

An imSensorList object (data.table) with all sensors collected for the respondent of interest.

Examples

## Not run: 
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
respondents <- imotionsApi::getRespondents(study)
sensors <- imotionsApi::getRespondentSensors(study, respondents[1, ])

## End(Not run)

[Package imotionsApi version 2.0.0 Index]