getRespondents {imotionsApi}	R Documentation

Get respondents from a study.

Description

Generic getRespondents function that take as parameter a study object and optionally a stimulus and/or segment object.

Usage

getRespondents(study, stimulus = NULL, segment = NULL)

Arguments

study
An imStudy object as returned from imStudy.

stimulus
Optional - An imStimulus object as returned from getStimuli.

segment
Optional - An imSegment object as returned from getSegments.

Value

An imRespondentList object (data.table) with all respondents of interest.

Examples

## Not run: 
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
stimuli <- imotionsApi::getStimuli(study)
segments <- imotionsApi::getSegments(study)

## Get all respondents in the study
respondents <- imotionsApi::getRespondents(study)

## Get all respondents exposed to a specific stimulus
respondents <- imotionsApi::getRespondents(study, stimulus = stimuli[1, ])

## Get all respondents in a specific segment
respondents <- imotionsApi::getRespondents(study, segment = segments[1, ])

## Get all respondents in a specific segment exposed to a specific stimulus
respondents <- imotionsApi::getRespondents(study, stimulus = stimuli[1, ], segment = segments[1, ])

## End(Not run)

[Package imotionsApi version 2.0.0 Index]