getRespondents.imSegment {imotionsApi}	R Documentation

Get respondents for a specific segment.

Description

S3 method to get all respondents for a specific segment.

Usage

## S3 method for class 'imSegment'
getRespondents(study, stimulus = NULL, segment)

Arguments

study
An imStudy object as returned from imStudy.

stimulus
Optional - An imStimulus object as returned from getStimuli.

segment
An imSegment object as returned from getSegments for which you would like to get the respondents.

Value

An imRespondentList object (data.table) listing all respondents in the segment specified.

[Package imotionsApi version 2.0.0 Index]