getRespondents.imStimulus {imotionsApi}	R Documentation

Get respondents for a specific stimulus.

Description

S3 method to get all respondents for a specific stimulus.

Usage

## S3 method for class 'imStimulus'
getRespondents(study, stimulus, ...)

Arguments

study
An imStudy object as returned from imStudy.

stimulus
An imStimulus object as returned from getStimuli for which you would like to get the respondents.

...
Further arguments passed from the getRespondents method.

Value

An imRespondentList object (data.table) with all respondents exposed to a specific stimulus.

[Package imotionsApi version 2.0.0 Index]