getSensorData {imotionsApi}	R Documentation

Download data corresponding to a specific sensor (signals/metrics).

Description

Available sensors in your study can be listed using the getRespondentSensors.

Usage

getSensorData(study, sensor, signalsName = NULL, intervals = NULL)

Arguments

study
An imStudy object as returned from imStudy.

sensor
An imSensor object as returned from getRespondentSensors.

signalsName
Optional: A vector of specific signals name you would like to return.

intervals
Optional: An imInterval or imIntervalList object with start/end of data to subset as given by getRespondentIntervals.

Details

Signals always have a "Timestamp" column and are unique to a given respondent and a given sensor source. Metrics are stored as a special sensor, also specific to a given respondent.

Value

An imData object (data.table) containing the signals (imSignals) or metrics (imMetrics) for the sensor of interest.

Examples

## Not run: 
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
respondents <- imotionsApi::getRespondents(study)
sensors <- imotionsApi::getRespondentSensors(study, respondents[1, ])
data <- imotionsApi::getSensorData(study, sensors[1, ])

## End(Not run)

[Package imotionsApi version 2.0.0 Index]