getStudySpecificUrl {imotionsApi}	R Documentation

Based on the kind of connection, return the path/url to a specific study object.

Description

Based on the kind of connection, return the path/url to a specific study object.

Usage

getStudySpecificUrl(study)

Arguments

study
An imStudy object as returned from imStudy.

[Package imotionsApi version 2.0.0 Index]