getStudyUrlById {imotionsApi}	R Documentation

Return the path/url to a specific study based of its studyId.

Description

Return the path/url to a specific study based of its studyId.

Usage

getStudyUrlById(connection, studyId)

Arguments

connection
An imConnection object as returned from imConnection.

studyId
The id of the study of interest.

[Package imotionsApi version 2.0.0 Index]