getUploadSensorsUrl {imotionsApi}	R Documentation

Generic getUploadSensorsUrl function that take as parameter a study object and a respondent/segment object. Return the path/url to upload a sensor to this respondent/segment object.

Description

Generic getUploadSensorsUrl function that take as parameter a study object and a respondent/segment object. Return the path/url to upload a sensor to this respondent/segment object.

Usage

getUploadSensorsUrl(study, imObject, stimulus = NULL)

Arguments

study
An imStudy object as returned from imStudy.

imObject
An imRespondent or imSegment object of interest.

stimulus
For segment object: an imStimulus object as returned from getStimuli.

[Package imotionsApi version 2.0.0 Index]