getUploadSensorsUrl.imRespondent {imotionsApi}	R Documentation

getUploadSensorsUrl method to get the path/url to upload a sensor to this respondent object.

Description

getUploadSensorsUrl method to get the path/url to upload a sensor to this respondent object.

Usage

## S3 method for class 'imRespondent'
getUploadSensorsUrl(study, imObject, ...)

Arguments

study
An imStudy object as returned from imStudy.

imObject
An imRespondent or imSegment object of interest.

[Package imotionsApi version 2.0.0 Index]