imConnection {imotionsApi}	R Documentation

Create a connection with the iMotions API.

Description

Tokens can be found on a study's R Analysis page or given directly by the iMotions software

Usage

imConnection(token, baseUrl = "https://my.imotions.com/api")

Arguments

token
The token to be used for authentication.

baseUrl
The iMotions server to connect to, normally you do not need to change the default.

Value

An imConnection object to be passed to other methods.

Examples

## Not run: 
myToken <- "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
connection <- imotionsApi::imConnection(myToken)

## End(Not run)

[Package imotionsApi version 2.0.0 Index]