imStudy {imotionsApi}	R Documentation

Load an iMotions study by id.

Description

Retrieves detailed information about a study including stimuli, respondents, segments, and more.

Usage

imStudy(connection, studyId)

Arguments

connection
An imConnection object as returned from imConnection.

studyId
The id of the study you would like to retrieve.

Details

Available studies can be found with listStudies.

Value

An imStudy object to be passed to other methods.

Examples

## Not run: 
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
studyId <- studies$id[1]
study <- imotionsApi::imStudy(connection, studyId)

## End(Not run)

[Package imotionsApi version 2.0.0 Index]