listLoadedStudies {imotionsApi}	R Documentation

List studies that have been loaded in the current session.

Description

More detailed information about a study can be retrieved using imStudy.

Usage

listLoadedStudies()

Value

A list of the studies that have been loaded in the current session.

Examples

## Not run: 
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
studyId <- studies$id[1]
study <- imotionsApi::imStudy(connection, studyId)
listLoadedStudies()

## End(Not run)

[Package imotionsApi version 2.0.0 Index]