privateUploadSignals {imotionsApi}	R Documentation

Upload signals to a given respondent/segment.

Description

Upload signals to a given respondent/segment.

Usage

privateUploadSignals(
  params,
  study,
  data,
  target,
  sensorName,
  scriptName,
  metadata = NULL
)

Arguments

params
The list of parameters provided to the script - specific parameters value will be stored as metadata.

study
An imStudy object as returned from imStudy.

data
A data.table containing the signals/metrics to upload (imData object are also accepted).

target
The target respondent for the sensor (an imRespondent object as returned from getRespondents).

sensorName
The name of the new sensor you would like to create.

scriptName
The name of the script used to produce these signals/metrics.

metadata
Optional - a data.table with metadata information. Column names will be converted to metadata headers and there must be a row corresponding to each data column.

[Package imotionsApi version 2.0.0 Index]