saveSignalsToFile {imotionsApi}	R Documentation

Create a temporary file with the data/metadata that needs to be uploaded.

Description

Create a temporary file with the data/metadata that needs to be uploaded.

Usage

saveSignalsToFile(params, data, sensorName, scriptName, metadata = NULL)

Arguments

params
The list of parameters provided to the script - specific parameters value will be stored as metadata.

data
A data.table containing the signals/metrics to upload (imData object are also accepted).

sensorName
The name of the new sensor you would like to create.

scriptName
The name of the script used to produce these signals/metrics.

metadata
Optional - a data.table with metadata information. Column names will be converted to metadata headers and there must be a row corresponding to each data column.

[Package imotionsApi version 2.0.0 Index]