stopOnHttpError {imotionsApi}	R Documentation

Analyse a status code from an HTTP response and stop if an error was encountered.

Description

Analyse a status code from an HTTP response and stop if an error was encountered.

Usage

stopOnHttpError(response, message = NULL)

Arguments

response
An HTTP response.

message
Optional: a short message indicating which steps are getting performed to get a more indicative error message.

Value

The last HTTP response.

[Package imotionsApi version 2.0.0 Index]