stopSubjectNotApplicable {imotionsApi}	R Documentation

Indicate that R markdown knitting should be stopped since the current stimulus/respondent/segment is not applicable for this notebook.

Description

Indicate that R markdown knitting should be stopped since the current stimulus/respondent/segment is not applicable for this notebook.

Usage

stopSubjectNotApplicable(message = NULL)

Arguments

message
Optional - a message explaining why this respondent was not applicable.

Examples

## Not run: 
stopSubjectNotApplicable("This respondent has no data.")

## End(Not run)

[Package imotionsApi version 2.0.0 Index]