unloadStudies {imotionsApi}	R Documentation

Remove studies that have been loaded in the current session.

Description

This function can be used when new changes have been made to already loaded studies.

Usage

unloadStudies()

Value

"Studies successfully removed from the current session" if success.

Examples

## Not run: 
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
studyId <- studies$id[1]
study <- imotionsApi::imStudy(connection, studyId)
unloadStudies()

## End(Not run)

[Package imotionsApi version 2.0.0 Index]