uploadSensorData {imotionsApi}	R Documentation

Create a new sensor for a specific respondent in a study.

Description

Either signals (with a Timestamp column) or metrics (single values, with no Timestamp) can be uploaded. After processing, the sensor can then be viewed and exported locally through the iMotions Desktop.

Usage

uploadSensorData(
  params,
  study,
  data,
  target,
  sensorName,
  scriptName,
  metadata = NULL
)

Arguments

params
The list of parameters provided to the script - specific parameters value will be stored as metadata.

study
An imStudy object as returned from imStudy.

data
A data.table containing the signals/metrics to upload (imData object are also accepted).

target
The target respondent for the sensor (an imRespondent object as returned from getRespondents).

sensorName
The name of the new sensor you would like to create.

scriptName
The name of the script used to produce these signals/metrics.

metadata
Optional - a data.table with metadata information. Column names will be converted to metadata headers and there must be a row corresponding to each data column.

Details

Params required field are "iMotionsVersion" and "flowName"

Examples

## Not run: 
connection <- imotionsApi::imConnection("xxxxxxxx")
studies <- imotionsApi::listStudies(connection)
study <- imotionsApi::imStudy(connection, studies$id[1])
respondents <- imotionsApi::getRespondents(study)
data <- data.frame("Timestamp" = seq(1:100), "Thresholded value" = rep(0, 100))
params <- list("iMotionsVersion" = 8, "flowName" = "Test")
uploadSensorData(params, study, data, respondents[1, ], "New sensor", "Example Script")

## End(Not run)

[Package imotionsApi version 2.0.0 Index]